requirements.txt
streamlit
openai
python-dotenv
cryptography
pandas
plotly
fpdf2
